@@ -1,7 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <!-- <LangVersion>latest</LangVersion> -->
   </PropertyGroup>
+  <ItemGroup>
+    <Folder Include="2021/Day01/" />
+  </ItemGroup>
 </Project>
